remove cookbook from docs and added to cookbook docset

+++ arango_dash.py
@@ -1,36 +1,59 @@
 #!/usr/local/bin/python
-
+import getopt
 import os
 import re
 import sqlite3
+from sys import argv
+import sys
 from bs4 import BeautifulSoup
 
-db = sqlite3.connect('ArangoDB.docset/Contents/Resources/docSet.dsidx')
-cur = db.cursor()
 
-folder = 'docs.arangodb.com'
+def main(arg):
+    try:
+        opts, args = getopt.getopt(arg, "ht:", ["docType="])
+    except getopt.GetoptError:
+        print "arango_dash.py -t <docType>"
+        sys.exit(2)
 
-try:
-    cur.execute('DROP TABLE searchIndex;')
-except:
-    pass
-cur.execute('CREATE TABLE searchIndex(id INTEGER PRIMARY KEY, name TEXT, type TEXT, path TEXT);')
-cur.execute('CREATE UNIQUE INDEX anchor ON searchIndex (name, type, path);')
+    for opt, ag in args:
+        if opt == '-h':
+            print "arango_dash.py -t <docType>"
+            sys.exit()
+        elif opt in ("-t", "-docType"):
 
-docpath = 'ArangoDB.docset/Contents/Resources/Documents/' + folder
+            docset_type = {'docs': 'ArangoDB', 'cookbook': 'ArangoDB-Cookbook'}[arg]
 
-page = open(os.path.join(docpath, 'index.html')).read()
-soup = BeautifulSoup(page)
+            db = sqlite3.connect(docset_type + '.docset/Contents/Resources/docSet.dsidx')
+            cur = db.cursor()
 
-any = re.compile('.*')
-for tag in soup.find_all('a', {'href': any}):
-    name = re.sub("\n\s+", " ", tag.text.strip())
-    if len(name) > 0:
-        path = re.sub("(\./)", "", tag.attrs['href'])
-        path = folder + "/" + path
-        if path.split('#')[0] not in ('index.html', 'biblio.html', 'bookindex.html'):
-            cur.execute('INSERT OR IGNORE INTO searchIndex(name, type, path) VALUES (?,?,?)', (name, 'Guide', path))
-            print 'name: %s, path: %s' % (name, path)
+            folder = 'docs.arangodb.com'
 
-db.commit()
-db.close()
+            try:
+                cur.execute('DROP TABLE searchIndex;')
+            except:
+                pass
+            cur.execute('CREATE TABLE searchIndex(id INTEGER PRIMARY KEY, name TEXT, type TEXT, path TEXT);')
+            cur.execute('CREATE UNIQUE INDEX anchor ON searchIndex (name, type, path);')
+
+            docpath = docset_type + '.docset/Contents/Resources/Documents/' + folder
+
+            page = open(os.path.join(docpath, 'index.html')).read()
+            soup = BeautifulSoup(page)
+
+            any = re.compile('.*')
+            for tag in soup.find_all('a', {'href': any}):
+                name = re.sub("\n\s+", " ", tag.text.strip())
+                if len(name) > 0:
+                    path = re.sub("(\./)", "", tag.attrs['href'])
+                    path = folder + "/" + path
+                    if path.split('#')[0] not in ('index.html', 'biblio.html', 'bookindex.html'):
+                        cur.execute('INSERT OR IGNORE INTO searchIndex(name, type, path) VALUES (?,?,?)',
+                                    (name, 'Guide', path))
+                        print 'name: %s, path: %s' % (name, path)
+
+            db.commit()
+            db.close()
+
+
+if __name__ == "__main__":
+    main(argv[1:])
